Update repair spell names
