+++ src/voodoo/spellbook/spells/index.ts
@@ -8,13 +8,13 @@ export { craftFlask } from './craftFlask';
 export { flaskOfEndlessTeleportation } from './flaskOfEndlessTeleportation';
 export { flaskOfEndlessWater } from './flaskOfEndlessWater';
 export { frostBolt } from './frostBolt';
-export { repairCopperWeapon } from './repairCopperWeapon';
-export { repairElectrumWeapon } from './repairElectrumWeapon';
-export { repairGoldWeapon } from './repairGoldWeapon';
-export { repairIronWeapon } from './repairIronWeapon';
-export { repairMythrilWeapon } from './repairMythrilWeapon';
-export { repairPalladiumWeapon } from './repairPalladiumWeapon';
-export { repairRedIronWeapon } from './repairRedIronWeapon';
-export { repairSilverWeapon } from './repairSilverWeapon';
-export { repairValyanWeapon } from './repairValyanWeapon';
-export { repairViridiumWeapon } from './repairViridiumWeapon';
+export { repairCopperDevice } from './repairCopperDevice';
+export { repairElectrumDevice } from './repairElectrumDevice';
+export { repairGoldDevice } from './repairGoldDevice';
+export { repairIronDevice } from './repairIronDevice';
+export { repairMythrilDevice } from './repairMythrilDevice';
+export { repairPalladiumDevice } from './repairPalladiumDevice';
+export { repairRedIronDevice } from './repairRedIronDevice';
+export { repairSilverDevice } from './repairSilverDevice';
+export { repairValyanDevice } from './repairValyanDevice';
+export { repairViridiumDevice } from './repairViridiumDevice';
